// src/Models/emp-agent.model.ts
export class Agent {

    agentId: string = '';
    agentFirstName: string = '';
    agentLastName: string = '';
    phoneNumber: number = 0;
    email: string = '';
    qualification: string = '';
    commission: string = '';
    password: string = '';
    userId: number = 0;

    userName: string = '';
    status: boolean = true;
}